pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "diffusion_policy"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Best_bachelor_4_ever", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clip",
    "mani-skill>=3.0.0b22",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "torch==2.9.1",
    "torchvision>=0.24.1",
    "typer==0.15.1",
    "ultralytics>=8.4.11",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mypy==1.19.1",
    "mkdocs-material==9.7.1",
    "mkdocstrings-python==2.0.1",
]


[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  { index = "pytorch-cpu", marker = "sys_platform != 'linux' and sys_platform != 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  { index = "pytorch-cpu", marker = "sys_platform != 'linux' and sys_platform != 'win32'" },
]
clip = { git = "https://github.com/openai/CLIP.git" }

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
